Move to GrainExtensions class
