+++ src/Orleans/Orleans.csproj
@@ -95,7 +95,6 @@
     <Compile Include="Core\IInternalClusterClient.cs" />
     <Compile Include="Core\IClusterClient.cs" />
     <Compile Include="Core\ClusterClient.cs" />
-    <Compile Include="Runtime\GrainMethodInvocationExtensions.cs" />
     <Compile Include="Runtime\GrainReferenceRuntime.cs" />
     <Compile Include="Serialization\GrainReferenceSerializer.cs" />
     <Compile Include="Providers\ILeaseProvider.cs" />
